Total overhaul, with proper delineation of roles between objects/functions.
This includes moving a lot of the really low-level sequence manipulation
and parsing stuff to a "sequtils" module. Chiefly, it sees the main jobs of
the old PySplicer being chopped into different objects.

These objects are separated into modules, such as DNAMapper for objects that
are used to identify areas of a DNA sequence that are not wanted (this now
includes rudimentary secondary structure detection), Translators for the
wrapper objects around codon frequency tables and translation tables,
CodonSplicer, which assembles a sequence with minimal undesired subsequences
or structures by splicing a set of candidates together, and CodonJuggler, which
uses mapping information to find alternative subsequences without the mapped
structures etc.

Also, the old "operation" code that was crammed (along with everything else)
into one file is now in a script in the "bin" directory, and will be installed
to a path directory if pysplicer is installed using pip-3.2 or by using
"python3 setup.py install".

If you are using "distribute" rather than the built-in distutils/setuptools
that comes with python, you have the additional option of using
"python3 setup.py develop" to perform a pseudo-installation of PySplicer, which
adds links to the module to python's path and the system path, but leaves all
the files in the directory they were downloaded to. This may be useful if you
want to use git to pull updates as they occur without requiring a full
reinstallation.

If you were using the old PySplicer from the Github repo, please upgrade ASAP,
as it has since been identified to have several critical bugs in the mapping
code. The Gitorious Repo (which is the primary repo) didn't have these bugs.