Added a bunch of features, refactored others.
Added enzyme_lib, which includes a REBASE-derived dictionary of enzymes and
several functions for parsing this data. Added a command-line option to the
main pysplicer script to handle this new feature, allowing terminal
specification of unwanted enzymes.

Added a feature to enrich codons in a codon table that have more of a certain
base, which can be used for adenine enrichment to help reduce secondary
structure formation (by avoiding promiscuous "U" bases and stronger pairing
"G" or "C" bases).