Added ViennaRNA wrapping module and integrated into PySplicer core script. Poorly tested so far.
Using ViennaRNA rather than my own hacked-together hairpin mapper will make secondary structure
avoidance far more robust and predictable. This in turn should help make more stable and Ribosome-
accessible mRNA and improve protein yields or at least translation predictability.

This new option is not default behaviour because it requires the ViennaRNA package to be installed
to the system's path. Ideally, the algorithms used in ViennaRNA could be ported to Python for a
slower but more portable pure-python fallback, but I don't have the time to do this.