Overhauled RNAFold usage, altered terminal options. More evidence based, less witchcraft.
Based on research showing secondary structure is mostly problematic only in early sequence,
and on research suggesting that early pausing in translation close to the RBS can actually
improve translation rates/efficiency, have reworked the code to support a maximum codon
frequency in this region and to restrict the very-costly RNA structure prediction operation
to this region also.

In verbose mode, the free energy of the early region (termed "avoid-ngg-span" in args) will
be printed to terminal.